requirements.txt
Flask
flask-cors
captcha
Werkzeug
patool
py7zr
Flask-SQLAlchemy
db-sqlite3
